@@ -7,7 +7,7 @@ description: An application that manages and receives alerts from ActingWeb acto
 # Both the version and the builder number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
-version: 1.3.3+7
+version: 1.3.4
 
 environment:
   sdk: ">2.1.0 <3.0.0"
@@ -17,29 +17,30 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_appauth: ^0.9.2+5
-  cupertino_icons: ^1.0.0
+  intl: ^0.17.0
+  flutter_appauth: ^0.9.2+6
+  cupertino_icons: ^1.0.2
   meta: ^1.1.8
-  provider: ^4.3.2+2
-  shared_preferences: ^0.5.12
-  google_maps_flutter: ^0.5.33 # Locking down to be <1.0.0 with Flutter 1.22 requirement
-  geolocator: ^5.3.1  # Latest with breaking changes: ^6.0.0+4
-  # geocoding: ^1.0.4+1 # Needs to be used for geolocator > 6.0.0
-  firebase_analytics: ^6.0.1
-  firebase_crashlytics: ^0.2.1
-  firebase_messaging: ^7.0.2
-  flushbar: ^1.10.4
+  provider: ^5.0.0
+  shared_preferences: ^2.0.3
+  google_maps_flutter: ^1.2.0
+  google_maps_flutter_web: ^0.1.1
+  geolocator: ^6.1.13
+  geocoding: ^1.0.5
+  firebase_core: ^0.7.0
+  firebase_analytics: ^7.0.1
+  firebase_crashlytics: ^0.4.0+1
+  firebase_messaging: ^8.0.0-dev.15
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  test: any
+  test: ^1.15.7
   flutter_launcher_icons: ^0.8.1
-  intl_translation: ^0.17.10+1
   convert: any
-  mockito: ^4.1.1 #Lock down as 4.1.2 seems to rely on a null-safe lib
+  mockito: ^4.1.3
 
 flutter_icons:
   android: "launcher_icon"
