@@ -66,7 +66,7 @@ class CustomDialog extends StatelessWidget {
               SizedBox(height: 24.0),
               Align(
                   alignment: Alignment.bottomRight,
-                  child: FlatButton(
+                  child: TextButton(
                     onPressed: () {
                       Navigator.of(context).pop(); // To close the dialog
                     },
