@@ -9,6 +9,8 @@ import 'intl/messages_all.dart';
 // **************************************************************************
 
 // ignore_for_file: non_constant_identifier_names, lines_longer_than_80_chars
+// ignore_for_file: join_return_with_assignment, prefer_final_in_for_each
+// ignore_for_file: avoid_redundant_argument_values
 
 class S {
   S();
@@ -333,21 +335,21 @@ class S {
     );
   }
 
-  /// `Expires`
-  String get drawerButtomSheetExpires {
-    return Intl.message(
-      'Expires',
-      name: 'drawerButtomSheetExpires',
-      desc: '',
-      args: [],
-    );
-  }
-
   /// `Firebase messaging token`
   String get drawerButtomSheetFCMToken {
     return Intl.message(
       'Firebase messaging token',
       name: 'drawerButtomSheetFCMToken',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Expires`
+  String get drawerButtomSheetExpires {
+    return Intl.message(
+      'Expires',
+      name: 'drawerButtomSheetExpires',
       desc: '',
       args: [],
     );
