@@ -13,7 +13,8 @@ import 'package:first_app/mock/mock_appauth.dart';
 import 'package:first_app/mock/mock_geolocator.dart';
 
 void main() async {
-  AppStateModel appState;
+  // Get an empty AppState
+  AppStateModel appState = new AppStateModel();
 
   // ignore: missing_return
   Future<String> dataHandler(String msg) async {
@@ -48,12 +49,18 @@ void main() async {
   // This line enables the extension.
   enableFlutterDriverExtension(handler: dataHandler);
 
+  // The driver does not like being enabled after the binding has been initialised,
+  // so we need to do it here
+  WidgetsFlutterBinding.ensureInitialized();
+
   // Get an instance so that globals are initialised
   var prefs = await SharedPreferences.getInstance();
+
   // We don't want any state we cannot control when testing
   prefs.clear();
   // Let's initialise the app state with the stored preferences
-  appState = new AppStateModel(prefs);
+  appState.prefs = prefs;
+  appState.setLocale(null);
 
   runApp(
     new MaterialApp(
